Moved PresentationFile related classes to a new package for cleanliness. Parser class and test fully commented. Ready for merging with Master.
