+++ app/src/main/java/com/presentech/handsup/presentationfile/Audio.java
@@ -1,22 +1,18 @@
-package com.presentech.handsup;
+package com.presentech.handsup.presentationfile;
 
 /**
- * Created by Alex on 25/02/2016.
+ * Created by Alex Butcher on 25/02/2016.
  */
-public class Video {
+public class Audio {
     String sourceFile;
     int startTime;
     int duration;
-    float xStart;
-    float yStart;
     boolean loop;
 
-    public Video(String sourceFile, int startTime, int duration, float xStart, float yStart, boolean loop) {
+    public Audio(String sourceFile, int startTime, int duration, boolean loop) {
         this.sourceFile = sourceFile;
         this.startTime = startTime;
         this.duration = duration;
-        this.xStart = xStart;
-        this.yStart = yStart;
         this.loop = loop;
     }
 
@@ -30,14 +26,6 @@ public class Video {
 
     public int getDuration() {
         return duration;
-    }
-
-    public float getxStart() {
-        return xStart;
-    }
-
-    public float getyStart() {
-        return yStart;
     }
 
     public boolean isLoop() {
