+++ app/src/main/java/com/presentech/handsup/presentationfile/Defaults.java
@@ -1,7 +1,7 @@
-package com.presentech.handsup;
+package com.presentech.handsup.presentationfile;
 
 /**
- * Created by Alex on 02/03/2016.
+ * Created by Alex Butcher on 02/03/2016.
  */
 public class Defaults {
