+++ app/src/main/java/com/presentech/handsup/presentationfile/Text.java
@@ -1,7 +1,7 @@
-package com.presentech.handsup;
+package com.presentech.handsup.presentationfile;
 
 /**
- * Created by Alex on 25/02/2016.
+ * Created by Alex Butcher on 25/02/2016.
  */
 public class Text {
     int startTime;
